Adding back original task
This is the original script from our initial published paper. We're providing it here for anyone interested in seeing the task as it was used, but we don't recommend anyone use it now. The scripts in the "Discounting tasks" directory are functionally equivalent to these original ones, but aspects of the scripts relating to reliability, ease of use, and datafile formatting have been improved.